move profile view
