+++ openstates/urls.py
@@ -11,6 +11,7 @@ from bulk.views import bulk_csv_list
 urlpatterns = [
     path("djadmin/", admin.site.urls),
     path("accounts/", include("allauth.urls")),
+    path("accounts/profile/", include("profiles.urls")),
     url("", include("geo.urls")),
     url(r"^api/v1/", include("v1.urls")),
     url(r"^api/confirm/$", ConfirmationView.as_view()),
@@ -23,7 +24,6 @@ urlpatterns = [
         ),
     ),
     url("", include("public.urls")),
-    url("profile/", include("profiles.urls")),
     url("", include("openstates.redirects")),
     path("bulk/csv/", bulk_csv_list),
 ]
