+++ public/templates/public/views/bills.html
@@ -147,7 +147,7 @@
                     {% if request.user.profile.feature_subscriptions %}
                     {% csrf_token %}
                     <input type="hidden" name="state" value="{{ state }}" />
-                    <button type="submit" class="button button--primary" formaction="/profile/add_search_sub/" formmethod="POST">Add Subscription</button>
+                    <button type="submit" class="button button--primary" formaction="/accounts/profile/add_search_sub/" formmethod="POST">Add Subscription</button>
                     {% endif %}
                 </form>
             </div>
